All integral costs, fix issue with swapStar

+++ hgs_vrptw/include/Config.h
@@ -15,9 +15,8 @@ struct Config
 
     // This was the default until now, but with this value feasible
     // individuals often become infeasible during the local search in
-    // educate. With initialTimeWarpPenalty = 10,
-    // this does not happen.
-    double initialTimeWarpPenalty = 1.0;
+    // educate. With initialTimeWarpPenalty = 10, this does not happen.
+    int initialTimeWarpPenalty = 1;
 
     size_t nbPenaltyManagement = 100;  // manage penalties every # iterations
     double penaltyBooster = 2.0;       // special increase penalty if no feas
@@ -64,7 +63,7 @@ struct Config
     explicit Config(size_t nbIter = 20'000,
                     int timeLimit = INT_MAX,
                     int seed = 0,
-                    double initialTimeWarpPenalty = 1.0,
+                    int initialTimeWarpPenalty = 1,
                     size_t nbPenaltyManagement = 100,
                     double penaltyBooster = 2.,
                     double penaltyIncrease = 1.2,
